requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic>=2.5.0
qdrant-client==1.12.1
google-generativeai>=0.8.0
python-dotenv==1.0.1
openai==1.57.4
numpy>=1.26.0
python-multipart==0.0.20